split up payload
